What you've done is more dangerous
With the setOp thing, it's all on the main thread and the op is taken
away immediately afterwards so there is no possibility of them using the
elevated permission for anything other than worldguard. I even had a
finally clause.

What you have done, although it may not seem too dangerous, it still
gives players the opportunity to use their elevated permissions.