Add option to allow helpers to WorldEdit
There are a couple of differences between helpers and trusted members.
Helpers can only build and interact with a plot while the owner is online.
Trusted members can do anything, including WorldEdit, at any time.

The distinction between online and offline access is important, and
helpers vs trusted makes sense for that, however, it can be desirable to
allow WorldEdit access to helpers. This option would allow helpers to use
WorldEdit only when the plot owner is online.