Improvement: Code refactoring (FLAG:misc-interact, frame-interact) (EVENT:interact-misc)
Final goal make the code more readable and clear.

The improvement is located here PlayerEvents_1_8.onInteract(PlayerInteractAtEntityEvent e)
Added in commit: 93c4854

Step 1:
Suggestion remove the special handling for AmorStand from the PlayerEvents_1_8.class and move it to the PlayerEvent.class

Step 2:
Refactoring the misc-interact flag and the interact-misc event.
After understanding the current code i came to the point. That the moving the handling for AmorStand to PlayerEvents makes most of the code obsolete since PlayerEvents.class handles all that stuff.

Step 3:
Remove unused code
The case INTERACT_MISC in EventUtil is unused -> no one set at any time the EventType for that
Remove the EventType INTERACT_MISC itself -> not used anymore

Step 4:
Add a new flag frame-interact for item frame this is for player that want seperate the interact between
ItemFrame and AmorStand