Move entity cap verification of armor stands & end crystals into player interact listener
When cancelled in the creature spawn event and placed by a player, the player will lose the armor stand / end crystal item from their inventory. This PR aims to fix this by moving the verification whether the entity should be cancelled up to the spawn egg right click, this way the item is not touched.

I tried my best to verify this is not breaking anything; I'm not sure if the original code in EntitySpawnListener should be kept? Reviews appreciated.

Signed-off-by: FlorianMichael <florian.michael07@gmail.com>